// client/types/Product.ts
export type ResponseProductType = {
  id: number;
  name: string;
  description: string;
  price: number;
  pack_size: number;
  brand: string;
  colour: string;
  img_url: string;
};

export type ProductType = {
  id: number;
  name: string;
  description: string;
  price: number;
  packSize: number;
  brand: string;
  colour: string;
  imgUrl: string;
};